move all Explorer-related files to src/components/Explorer
