+++ src/components/Explorer/QueryResultPage.jsx
@@ -1,8 +1,8 @@
 import React from 'react';
 import { withFauxDOM } from 'react-faux-dom'
 import * as d3 from 'd3';
-import { drawQueryResults } from '../../SDK/drawQueryResults.js';
-import { drawRunResults } from '../../SDK/drawRunResults.js';
+import { drawQueryResults } from './SDK/drawQueryResults.js';
+import { drawRunResults } from './SDK/drawRunResults.js';
 import { ExternalLink } from "../../helpers/common";
 
 const QueryResult = (props) => {
