+++ src/components/Explorer/SDK/d3RangeSlider.js
